2655 illumos should be lint 12.3 free Reviewed by: Garrett D'Amore <garrett@damore.org> Reviewed by: Andrew Stormont <Andrew.Stormont@nexenta.com> Approved by: Richard Lowe <richlowe@richlowe.net>

+++ usr/src/cmd/mpathadm/mpathadm.c
@@ -3144,7 +3144,7 @@ displayWideArray(MP_WCHAR *arrayToDisplay, int arraySize)
 
 	for (i = 0; i < numChars; i++) {
 		if (L'\0' != arrayToDisplay[i]) {
-			(void) fprintf(stdout, "%wc", arrayToDisplay[i]);
+			(void) fprintf(stdout, "%wc", (int)arrayToDisplay[i]);
 		}
 	}
 }
